feat: split main and cli
