+++ crates/prover/src/bin/main.rs
@@ -1,45 +1,5 @@
-use clap::{Parser, Subcommand};
-use tracing::Level;
-
-use mojave_chain_utils::{logging::init_logging, prover_options::ProverOpts};
-
-use mojave_prover::ProverServer;
-
-#[derive(Parser)]
-#[command(
-    name = "mojave-prover",
-    author,
-    version,
-    about = "Mojave Prover service for the Mojave network",
-    arg_required_else_help = true
-)]
-pub struct Cli {
-    #[arg(
-      long = "log.level",
-      default_value_t = Level::INFO,
-      value_name = "LOG_LEVEL",
-      help = "The verbosity level used for logs.",
-      long_help = "Possible values: info, debug, trace, warn, error",
-      help_heading = "Node options")]
-    pub log_level: Level,
-    #[command(subcommand)]
-    pub command: Command,
-}
-
-impl Cli {
-    pub fn run() -> Self {
-        Self::parse()
-    }
-}
-
-#[derive(Subcommand)]
-pub enum Command {
-    #[command(name = "init", about = "Run the prover")]
-    Init {
-        #[command(flatten)]
-        prover_options: ProverOpts,
-    },
-}
+use mojave_chain_utils::logging::init_logging;
+use mojave_prover::{Cli, Command, ProverServer};
 
 #[tokio::main]
 async fn main() {
