+++ crates/prover/src/cli.rs
@@ -1,0 +1,40 @@
+use clap::{Parser, Subcommand};
+use tracing::Level;
+
+use mojave_chain_utils::prover_options::ProverOpts;
+
+#[derive(Parser)]
+#[command(
+    name = "mojave-prover",
+    author,
+    version,
+    about = "Mojave Prover service for the Mojave network",
+    arg_required_else_help = true
+)]
+pub struct Cli {
+    #[arg(
+      long = "log.level",
+      default_value_t = Level::INFO,
+      value_name = "LOG_LEVEL",
+      help = "The verbosity level used for logs.",
+      long_help = "Possible values: info, debug, trace, warn, error",
+      help_heading = "Prover options")]
+    pub log_level: Level,
+    #[command(subcommand)]
+    pub command: Command,
+}
+
+impl Cli {
+    pub fn run() -> Self {
+        Self::parse()
+    }
+}
+
+#[derive(Subcommand)]
+pub enum Command {
+    #[command(name = "init", about = "Run the prover")]
+    Init {
+        #[command(flatten)]
+        prover_options: ProverOpts,
+    },
+}
