+++ crates/prover/src/lib.rs
@@ -1,4 +1,4 @@
-// mod cli;
+mod cli;
 #[cfg(feature = "client")]
 mod client;
 mod message;
@@ -6,6 +6,7 @@ mod message;
 mod server;
 mod types;
 
+pub use cli::*;
 #[cfg(feature = "client")]
 pub use client::{ProverClient, ProverClientError};
 #[cfg(feature = "server")]
